ref: Metrics around symbolication task spawning (#703)
This experiments with metrics targetted around the symbolication task
spawning.  It does a manual histogram-timer first spawn duration as we
have been using in various places before.

It also experiments with adding tokio-metrics to measure the same
thing in more detail.  The reason to do both is to be able to compare
and evaluate tokio-metrics a bit.